pm_assistant/requirements.txt
google-adk ==1.28.0
psycopg2-binary
python-dotenv